+++ common/src/main/java/me/pandadev/fallingtrees/mixin/renderer/BlockRenderDispatcherAccessor.java
@@ -1,0 +1,12 @@
+package me.pandadev.fallingtrees.mixin.renderer;
+
+import net.minecraft.client.color.block.BlockColors;
+import net.minecraft.client.renderer.block.BlockRenderDispatcher;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+@Mixin(BlockRenderDispatcher.class)
+public interface BlockRenderDispatcherAccessor {
+	@Accessor
+	BlockColors getBlockColors();
+}