+++ common/src/main/java/me/pandadev/fallingtrees/tree/TreeUtils.java
@@ -1,14 +1,13 @@
 package me.pandadev.fallingtrees.tree;
 
 import me.pandadev.fallingtrees.FallingTrees;
-import me.pandadev.fallingtrees.FallingTreesConfig;
+import me.pandadev.fallingtrees.config.ClientConfig;
 import me.pandadev.fallingtrees.entity.TreeEntity;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.core.registries.BuiltInRegistries;
-import net.minecraft.sounds.SoundSource;
 import net.minecraft.stats.Stats;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.AxeItem;
@@ -19,37 +18,38 @@ import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.LeavesBlock;
 import net.minecraft.world.level.block.state.BlockState;
+import org.joml.Math;
 import org.joml.Vector3d;
 
 import java.util.*;
 
 public class TreeUtils {
 	public static boolean isLog(Block block) {
-		if (FallingTrees.serverConfig.blacklisted_log_blocks.contains(BuiltInRegistries.BLOCK.getKey(block).toString()))
+		if (FallingTrees.getServerConfig().blacklisted_log_blocks.contains(BuiltInRegistries.BLOCK.getKey(block).toString()))
 			return false;
-		if (FallingTrees.serverConfig.whitelisted_log_blocks.contains(BuiltInRegistries.BLOCK.getKey(block).toString()))
+		if (FallingTrees.getServerConfig().whitelisted_log_blocks.contains(BuiltInRegistries.BLOCK.getKey(block).toString()))
 			return true;
 
 		return block.defaultBlockState().getTags().anyMatch(blockTagKey ->
-				FallingTrees.serverConfig.whitelisted_log_block_tags.contains(blockTagKey.location().toString()));
+				FallingTrees.getServerConfig().whitelisted_log_block_tags.contains(blockTagKey.location().toString()));
 	}
 
 	public static boolean isLeaves(Block block) {
-		if (FallingTrees.serverConfig.blacklisted_leaves_blocks.contains(BuiltInRegistries.BLOCK.getKey(block).toString()))
+		if (FallingTrees.getServerConfig().blacklisted_leaves_blocks.contains(BuiltInRegistries.BLOCK.getKey(block).toString()))
 			return false;
-		if (FallingTrees.serverConfig.whitelisted_leaves_blocks.contains(BuiltInRegistries.BLOCK.getKey(block).toString()))
+		if (FallingTrees.getServerConfig().whitelisted_leaves_blocks.contains(BuiltInRegistries.BLOCK.getKey(block).toString()))
 			return true;
 
 		return block.defaultBlockState().getTags().anyMatch(blockTagKey ->
-				FallingTrees.serverConfig.whitelisted_leaves_block_tags.contains(blockTagKey.location().toString()));
+				FallingTrees.getServerConfig().whitelisted_leaves_block_tags.contains(blockTagKey.location().toString()));
 	}
 
 	@Environment(EnvType.CLIENT)
 	public static boolean isMiningOneBlock(Player player) {
-		if (FallingTrees.configHolder.getConfig().one_block_mining_method.equals(FallingTreesConfig.OneBlockMiningEnum.CROUCH)) {
+		if (FallingTrees.getClientConfig().one_block_mining_method.equals(ClientConfig.OneBlockMiningEnum.CROUCH)) {
 			return player.isCrouching();
 		}
-		return FallingTrees.configHolder.getConfig().is_mining_one_block;
+		return FallingTrees.getClientConfig().is_mining_one_block;
 	}
 
 //	public static boolean isDecorative(Block block) {
@@ -190,16 +190,17 @@ public class TreeUtils {
 
 				treeEntity.setRotationY((float) Math.atan2(player.getX() - position.x, player.getZ() - position.z));
 				level.addFreshEntity(treeEntity);
-				if (FallingTrees.serverConfig.sound_effect)
-					level.playSound(null, blockPos, FallingTrees.TREE_FALL.get(), SoundSource.BLOCKS,
-							0.25f*FallingTrees.serverConfig.sound_effect_volume, 1);
+//				if (FallingTrees.serverConfig.sound_effect)
+//					level.playSound(null, blockPos, FallingTrees.TREE_FALL.get(), SoundSource.BLOCKS,
+//							0.25f*FallingTrees.serverConfig.sound_effect_volume, 1);
 
 				int LogAmount = TreeUtils.getAmountOfLogs(treeBlocks);
 				if (usedItem.isDamageableItem()) {
-					usedItem.hurtAndBreak((int) (LogAmount * FallingTrees.serverConfig.item_damage_multiplier), player, player1 -> {});
+					usedItem.hurtAndBreak((int) (LogAmount * FallingTrees.getServerConfig().item_damage_multiplier), player, player1 -> {});
 				}
 
-				player.causeFoodExhaustion(0.005F * LogAmount * FallingTrees.serverConfig.food_exhaustion_multiplier);
+				player.causeFoodExhaustion(Math.min(0.005F * LogAmount * FallingTrees.getServerConfig().food_exhaustion_multiplier,
+						FallingTrees.getServerConfig().food_exhaustion_limit));
 
 				for (Map.Entry<BlockPos, BlockState> entry : treeBlocks.entrySet()) {
 					player.awardStat(Stats.BLOCK_MINED.get(entry.getValue().getBlock()));
@@ -211,9 +212,9 @@ public class TreeUtils {
 
 	@Environment(EnvType.CLIENT)
 	public static boolean shouldTreeFall(Player player) {
-		if (FallingTrees.serverConfig.tree_limit.only_fall_on_tool_use) {
-			return player.getMainHandItem().getItem() instanceof AxeItem && !(FallingTrees.serverConfig.allow_one_block_mining && TreeUtils.isMiningOneBlock(player));
+		if (FallingTrees.getServerConfig().tree_limit.only_fall_on_tool_use) {
+			return player.getMainHandItem().getItem() instanceof AxeItem && !(FallingTrees.getServerConfig().allow_one_block_mining && TreeUtils.isMiningOneBlock(player));
 		}
-		return !(FallingTrees.serverConfig.allow_one_block_mining && TreeUtils.isMiningOneBlock(player));
+		return !(FallingTrees.getServerConfig().allow_one_block_mining && TreeUtils.isMiningOneBlock(player));
 	}
 }