+++ fabric/src/main/java/me/pandadev/fallingtrees/fabric/ModMenuIntegration.java
@@ -2,7 +2,7 @@ package me.pandadev.fallingtrees.fabric;
 
 import com.terraformersmc.modmenu.api.ConfigScreenFactory;
 import com.terraformersmc.modmenu.api.ModMenuApi;
-import me.pandadev.fallingtrees.FallingTreesConfig;
+import me.pandadev.fallingtrees.config.ModConfig;
 import me.shedaniel.autoconfig.AutoConfig;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
@@ -11,6 +11,6 @@ import net.fabricmc.api.Environment;
 public class ModMenuIntegration implements ModMenuApi {
 	@Override
 	public ConfigScreenFactory<?> getModConfigScreenFactory() {
-		return screen -> AutoConfig.getConfigScreen(FallingTreesConfig.class, screen).get();
+		return screen -> AutoConfig.getConfigScreen(ModConfig.class, screen).get();
 	}
 }