+++ Changelog.md
@@ -1,5 +1,8 @@
 ## Version 0.6 Changelog
 * **Removed**
-  * Mining speed by size of tree.
+  * Mining speed by size of tree. (Will be added back later)
 * **Added**
-  * [HT's TreeChop](https://www.curseforge.com/minecraft/mc-mods/treechop) support.
+  * One block mining indicator
+  * Food exhaustion limit
+* **Changes**
+  * Updated Networking to fix crashes when breaking logs.